Crate iota_genesis_builder

Modules

genesis_build_effects

stardust

The stardust module incorporates all the logic necessary for parsing Stardust UTXOs from a full-snapshot file, and converting them to the appropriate genesis objects.

validator_info

Structs

Builder

Enums

SnapshotSource

SnapshotUrl

The URLs to download IOTA object snapshot.

Constants

GENESIS_BUILDER_PARAMETERS_FILE

IOTA_OBJECT_SNAPSHOT_URL

OBJECT_SNAPSHOT_FILE_PATH

Functions

generate_genesis_system_object

split_timelocks