Module iota_genesis_builder::stardust::migration

Modules

verification

The verification module contains the validation logic to make sure that the stardust outputs are correctly converted to the move objects.

Structs

Migration

The orchestrator of the migration process.

MigrationObjects

All the objects created during the migration.

Enums

MigrationTargetNetwork

The target network of the migration.

Constants

MIGRATION_PROTOCOL_VERSION

We fix the protocol version used in the migration.

PACKAGE_DEPS

The dependencies of the generated packages for native tokens.

Type Aliases

ExpirationTimestamp

An alias for representing the timestamp of a Timelock