iota_genesis_builder::stardust::migration

Module verification

The verification module contains the validation logic to make sure that the stardust outputs are correctly converted to the move objects.

Modules

alias

basic

created_objects

foundry

nft