iota_genesis_builder::stardust::native_token

Module package_data

The package_data module provides the NativeTokenPackageData struct, which encapsulates all the data necessary to build a Stardust native token package.

Structs

Irc30MetadataAlternative

NativeTokenModuleData

The NativeTokenModuleData struct encapsulates all the data necessary to build a Stardust native token module.

NativeTokenPackageData

The NativeTokenPackageData struct encapsulates all the data necessary to build a Stardust native token package.