iota_genesis_builder::stardust

Module process_outputs

Functions

is_participation_output

process_outputs_for_iota

Processes an iterator of outputs coming from a Hornet snapshot chaining some filters:

scale_amount_for_iota

Take an amount and scale it by a multiplier defined for the IOTA token.