iota_genesis_builder::stardust::process_outputs

Function process_outputs_for_iota

pub fn process_outputs_for_iota<'a>(
    target_milestone_timestamp: u32,
    swap_split_map: AddressSwapSplitMap,
    outputs: impl Iterator<Item = Result<(OutputHeader, Output)>> + 'a,
) -> impl Iterator<Item = Result<(OutputHeader, Output), Error>> + 'a

Processes an iterator of outputs coming from a Hornet snapshot chaining some filters:

the ScaleIotaAmountIterator scales balances of IOTA Tokens from micro to nano.

the UnlockedVestingIterator takes vesting outputs that can be unlocked and merges them into a unique basic output.

the ParticipationOutputFilter removes all features from the basic outputs with a participation tag.

the SwapSplitIterator performs the operation of SwapSplit given a map as input, i.e., for certain origin addresses it swaps the addressUC to a destination address and splits some amounts of tokens and/or timelocked tokens (this operation can be done for several destinations).