iota_genesis_builder::stardust::types::address_swap_map

Struct AddressSwapMap

pub struct AddressSwapMap { /* private fields */ }

Implementations

impl AddressSwapMap

pub fn destination_address( &self, origin_address: &IotaAddress, ) -> Option<IotaAddress>

Retrieves the destination address for a given origin address.

pub fn verify_all_addresses_swapped(&self) -> Result<()>

Verifies that all addresses have been swapped at least once. Returns an error if any address is not swapped.

pub fn stardust_to_iota_address_owner( &self, stardust_address: impl Into<StardustAddress>, ) -> Result<Owner>

Converts a StardustAddress to an Owner by first converting it to an IotaAddress and then checking against the swap map for potential address substitutions.

If the address exists in the swap map, it is swapped with the mapped destination address before being wrapped into an Owner.

pub fn swap_stardust_to_iota_address_owner( &mut self, stardust_address: impl Into<StardustAddress>, ) -> Result<Owner>

Converts a StardustAddress to an Owner by first converting it to an IotaAddress and then checking against the swap map for potential address substitutions.

If the address exists in the swap map, it is swapped with the mapped destination address before being wrapped into an Owner.

pub fn swap_stardust_to_iota_address( &mut self, stardust_address: impl Into<StardustAddress>, ) -> Result<IotaAddress>

Converts a StardustAddress to an IotaAddress and checks against the swap map for potential address substitutions.

If the address exists in the swap map, it is swapped with the mapped destination address before being returned as an IotaAddress.

pub fn from_csv(file_path: &str) -> Result<AddressSwapMap, Error>

Initializes an AddressSwapMap by reading address pairs from a CSV file.

The function expects the file to contain two columns: the origin address (first column) and the destination address (second column). These are parsed into a HashMap that maps origin addresses to tuples containing the destination address and a flag initialized to false.

Example CSV File

Origin,Destination
iota1qp8h9augeh6tk3uvlxqfapuwv93atv63eqkpru029p6sgvr49eufyz7katr,0xa12b4d6ec3f9a28437d5c8f3e96ba72d3c4e8f5ac98d17b1a3b8e9f2c71d4a3c
iota1qp7h2lkjhs6tk3uvlxqfjhlfw34atv63eqkpru356p6sgvr76eufyz1opkh,0x42d8c182eb1f3b2366d353eed4eb02a31d1d7982c0fd44683811d7036be3a85e

Comments are optional, and start with a # character. Only entries that start with this character are treated as comments.

Parameters

file_path: The relative path to the CSV file containing the address mappings.

Returns

An AddressSwapMap containing the parsed mappings.

Errors

Returns an error if the file cannot be found, read, or parsed correctly.

Returns an error if the origin or destination addresses cannot be parsed into an IotaAddress.