iota_genesis_builder::stardust::types::output::basic

Trait BasicOutputExt

pub trait BasicOutputExt {
    // Required methods
    fn new_from_stardust(
        header_object_id: ObjectID,
        output: &BasicOutput,
    ) -> Result<BasicOutput>;
    fn to_genesis_object(
        &self,
        owner: IotaAddress,
        protocol_config: &ProtocolConfig,
        tx_context: &TxContext,
        version: SequenceNumber,
        coin_type: &CoinType,
    ) -> Result<Object>;
    fn into_genesis_coin_object(
        self,
        owner: IotaAddress,
        protocol_config: &ProtocolConfig,
        tx_context: &TxContext,
        version: SequenceNumber,
        coin_type: &CoinType,
    ) -> Result<Object>;
    fn is_simple_coin(&self, target_milestone_timestamp_sec: u32) -> bool;
}

Extension trait for creating BasicOutput from Stardust types.

Required Methods

fn new_from_stardust( header_object_id: ObjectID, output: &BasicOutput, ) -> Result<BasicOutput>

Construct the basic output with an empty Bag using the Output Header ID and Stardust BasicOutput.

fn to_genesis_object( &self, owner: IotaAddress, protocol_config: &ProtocolConfig, tx_context: &TxContext, version: SequenceNumber, coin_type: &CoinType, ) -> Result<Object>

Creates a genesis object from this basic output.

fn into_genesis_coin_object( self, owner: IotaAddress, protocol_config: &ProtocolConfig, tx_context: &TxContext, version: SequenceNumber, coin_type: &CoinType, ) -> Result<Object>

Converts this basic output into a genesis coin object.

fn is_simple_coin(&self, target_milestone_timestamp_sec: u32) -> bool

Infer whether this object can resolve into a simple coin.

Returns true in particular when the given milestone timestamp is equal or past the unix timestamp in a present timelock and no other unlock condition or metadata, tag, sender feature is present.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.