iota_genesis_builder::stardust::types::output

Module unlock_conditions

Extension traits for creating unlock conditions from Stardust types during migration.

Structs

ExpirationUnlockCondition

Rust version of the stardust expiration unlock condition.

StorageDepositReturnUnlockCondition

Rust version of the stardust storage deposit return unlock condition.

TimelockUnlockCondition

Rust version of the stardust timelock unlock condition.

Traits

ExpirationUnlockConditionExt

Extension trait for creating ExpirationUnlockCondition from Stardust types.

StorageDepositReturnUnlockConditionExt

Extension trait for creating StorageDepositReturnUnlockCondition from Stardust types.

TimelockUnlockConditionExt

Extension trait for creating TimelockUnlockCondition from Stardust types.